Feature: Added index action for link controller
I need to think about how I want to do this, because if this link controller
is going to continue to exist, it will be nested within the box index, which
will be the real "index" page.

I almost think it would be easier to just have a the box model that has 7 fields,
name, 3 titles, and 3 links. That way I could use the put/patch HTTP request and
wouldn't have to add a new link and hide the old one. It makes sense because it's
editing rather than producing, but it is hard coding only 3 links, so it would
be difficult if you ever decide you want more than that.

But what is brevity box without the limit of 3 links?