Feature: Add a show action for the link controller
This almost certainly will end up being deleted, because it doesn't make sense
to have a page just to show one link. See my commit message on adding the link
index. More and more I'm thinkin it should just be the box controller that has
3 links in it and a name. Or maybe the name would be tied to the user model.
Is the user model even going to be complex enough to be separate?