Feature: Add temporary link to individual boxes
I put in a generic "Name of Person" link that goes to an individual box
using the id. This was also in connection with adding the edit method to the
boxes controller, becuase I needed a way to navigate to an individual box
to edit it (I don't want edit links on the index page).

Eventually this "Name of Person" string will be replaced by the person's name
who owns the box, and it will still link to the show page for the box.